Rename tmux directory to make it explicit
I need to run extra bit of code when I am using tmux on a remote server
to force tmux to update SSH_AUTH_SOCK. This code should not run on a
local tmux instance. For this I have created tmux-local and tmux-remote
so when I create a new remote or local server, I'd be forced to pick one
or the other.

I thought about adding the code in one tmux config in a conditional. For
that I need to be able to figure out if I am logging in via ssh or
localhost, and I am not sure how to do that yet.